ddev/pyproject.toml
[build-system]
requires = [
    "hatchling>=1.5.0",
]
build-backend = "hatchling.build"

[project]
name = "ddev"
description = "The Datadog Agent integration developer tool"
readme = "README.md"
requires-python = ">=3.8"
license = "BSD-3-Clause"
keywords = [
    "datadog",
    "agent",
    "integration",
]
authors = [
    { name = "Datadog", email = "packages@datadoghq.com" },
]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: BSD License",
    "Natural Language :: English",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3.8",
]
dependencies = [
    "datadog-checks-dev[cli]",
    "hatch>=1.3.1",
    "rich>=12.5.1",
]
dynamic = ["version"]

[project.scripts]
ddev = "ddev.cli:main"

[project.urls]
Source = "https://github.com/DataDog/integrations-core"

[tool.hatch.version]
path = "src/ddev/__about__.py"
